Add buildkit --secret for --mount=type=secret
One of the experimental syntaxes for the Dockerfile frontend is `RUN
--mount=type=secret` which when used in conjunction with `--secret` flags, can
mount secret files from the host at build time without including them in the
layers. This is useful for credentials needed to access restricted resources
such as private registries and repositories. Read more about it at
https://github.com/moby/buildkit/blob/master/frontend/dockerfile/docs/experimental.md#run---mounttypesecret

This change adds the ability to effectively pass in `--secret` flags by using
BUILDKIT_SECRET_* environment variables. The mapping is described in the README
changes.